Minor: Counsil Group data request and alerts
