+++ src/app/council-group/search/search-council-group.component.css
@@ -7,6 +7,10 @@
 .text-center {
   margin-top: 10%;
 }
+.bq-danger{
+  width: 100%;
+  height: 10%;
+}
 .select-style {
     border: 1px solid #ccc;
     width: 70%;
